core.excludesfile now depends on $HOME
This patch sets the core.excludesfile git property to
“$HOME/.gitignore.global”.  It doesn't look like git config files can
contain parameter references, though, so we have to be a bit clever.

First, we use a bash startup script to set this property at every login.
Bash will expand out $HOME for us, and we'll set the property to be the
fully expanded value.

As a result, though, we can't store .gitconfig itself in the repository,
since then we'd get spurious complaints from git-status all the time.
So instead, we put all of the static git properties in a .gitconfig.in
file, and add this to the repository.  The bash startup script, then,
copies this file over to the real .gitconfig, and then uses git-config
to set the dynamic properties.