Fix quoting in bash prompt so fix weird wrapping issues
The problem seemed to be related to the color codes in the git branch
portion of the prompt.  I've changed things so that the reference to the
color code variables ${pblue} and ${preset} are evaluated when PS1 is
defined; so these get turned into the appropriate “\[\e...” codes right
away.  The “$(__git_ps1 ...)” call, on the other hand, is wrapped in
single quotes so that bash evaluates this when calculating the prompt.
The end result is that the arguments to __git_ps1 contain the actual
color codes, and not the helpful parameter.  This, for some reason,
seems to eliminate the wrapping behavior.