Adding platform-specific install directories
Ideally any program that I need will be available from the OS's
package manager — for the OSes that I use currently, either Ubuntu's
apt-get repository or Fink.  Not everything is, though.  So, I've
decided to track any extra programs I have to install in this git
repository.  So that this works seamlessly on different platforms,
I'll install each program into a platform-specific directory:

  ~/platforms/${OS}-${HOSTTYPE}

where OS is `uname -s` and HOSTTYPE is `uname -m`.  So on an Intel
Linux box, that would be ~/platforms/Linux-i686.

This patch makes this directory name available as PLATFORM_DIR.  It
also adds the platform-specific bin directory to my path, if that
directory exists.