need function now used with if statement
Previously, the need function would take in a list of commands to
execute on stdin; if the specified file/command exists, the commands
would be “eval”ed by the bash shell.  This meant that we had to use a
weird syntax inside the list of commands.

With this patch, the need function no longer executes the list of
commands itself; instead, it just returns a true (0) or false (1) status
code depending on whether the command exists.  This should be used with
a bash “if” statement to execute something when the command does exist.

This means the need is now very similar to have_program; the difference
between the two is that need prints out a warning message if the command
doesn't exist.  (And both now correctly return bash's notion of true
(0), which is backwards from most other programming languages...)