Set keys alias correctly, regardless of which SSH keys are installed
The keychain script now sets up my “keys” alias (which loads my SSH keys
into the ssh-agent process) to work regardless of which particular keys
are installed in my $HOME/.ssh directory.  It also defines the alias to
pass in the appropriate --dir and --host options to the keychain
command.