Only use __git_ps1 when present
We use the __git_ps1 function, which is part of the git bash-completion
script, to display the current git branch in the prompt.  If the
bash-completion support hasn't been installed, we would get an ugly
"command not found" error after every command.  This patch only updates
the prompt if the __git_ps1 function is actually present.